requirements.txt
duckduckgo-search
colorama
requests
beautifulsoup4
trafilatura
readchar
keyboard
curses-windows; sys_platform == 'win32'
tqdm
urllib3
openai>=1.0.0
anthropic>=0.7.0
tavily-python
exa-py
flask
python-dotenv